Use a better pattern for finding the gam path
sys.argv[0] is probably a more accepted and accurate pattern for doing
this task. It will always point at the executable that was run, whether
that's `gam.py` or `python -m gam`, both of which should be in the same
directory as the complimentary files that gam will be looking for -
oauth2.txt, oauth2service.txt, etc.